apps/api/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database (SQLite - no Docker needed)
sqlalchemy==2.0.27
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Image Processing
Pillow>=10.3.0

# Vector Operations (in-memory, no pgvector)
numpy>=2.0.0

# HTTP Clients
httpx==0.26.0

# Utilities
python-dotenv==1.0.1

# Development/Testing
pytest==7.4.4
pytest-cov==4.1.0
